services/qdrant_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
qdrant-client==1.9.2
openai==1.61.0
